refactor: Move once_by_key.go into lib/status
